Draft prof2 parser for profile.db file

+++ edu.rice.cs.hpcdata/src/edu/rice/cs/hpcdata/db/version4/MetricValueCollection3.java
@@ -66,7 +66,7 @@ public class MetricValueCollection3 implements IMetricValueCollection
 				values = new IntObjectHashMap<MetricValue>(sparseValues.size()); //new HashMap<Integer, MetricValue>(sparseValues.size());
 				
 				for (MetricValueSparse mvs: sparseValues) {
-					float value = mvs.getValue();
+					float value = (float) mvs.getValue();
 					float annotationValue = 1.0f;
 					
 					// compute the percent annotation
